# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AssessmentRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "default": "mock",
            "enum": [
              "mock",
              "live"
            ],
            "title": "Mode",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileIn"
          }
        },
        "required": [
          "profile"
        ],
        "title": "AssessmentRequest",
        "type": "object"
      },
      "AssessmentSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "daily_calorie_target": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Calorie Target"
          },
          "fitness_goal": {
            "title": "Fitness Goal",
            "type": "string"
          },
          "fitness_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fitness Level"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "injury_risk": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Injury Risk"
          },
          "mode": {
            "enum": [
              "mock",
              "live"
            ],
            "title": "Mode",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "safety_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Level"
          },
          "sessions_per_week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessions Per Week"
          },
          "simulated": {
            "title": "Simulated",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "complete",
              "partial",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          }
        },
        "required": [
          "id",
          "created_at",
          "parent_id",
          "mode",
          "simulated",
          "status",
          "user_name",
          "fitness_goal",
          "fitness_level",
          "injury_risk",
          "safety_level",
          "daily_calorie_target",
          "sessions_per_week"
        ],
        "title": "AssessmentSummary",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReplanRequest": {
        "properties": {
          "mode": {
            "default": "mock",
            "enum": [
              "mock",
              "live"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "title": "ReplanRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Personalised fitness assessments combining calibrated ML classifiers with an LLM planning workflow. Live generation is bring-your-own-key: send your Gemini key in `X-Gemini-Api-Key` (never stored or logged).",
    "title": "AetherFit AI API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assessments": {
      "get": {
        "operationId": "list_assessments_api_v1_assessments_get",
        "parameters": [
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssessmentSummary"
                  },
                  "title": "Response List Assessments Api V1 Assessments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List saved assessments for this device",
        "tags": [
          "assessments"
        ]
      },
      "post": {
        "operationId": "create_assessment_api_v1_assessments_post",
        "parameters": [
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          },
          {
            "description": "Bring-your-own Gemini API key. Used for this request only; never stored or logged.",
            "in": "header",
            "name": "x-gemini-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bring-your-own Gemini API key. Used for this request only; never stored or logged.",
              "title": "X-Gemini-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assessment"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or missing Gemini API key (live mode)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited (this API or the Gemini key's quota)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Generate an assessment",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/stream": {
      "post": {
        "operationId": "stream_assessment_api_v1_assessments_stream_post",
        "parameters": [
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          },
          {
            "description": "Bring-your-own Gemini API key. Used for this request only; never stored or logged.",
            "in": "header",
            "name": "x-gemini-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bring-your-own Gemini API key. Used for this request only; never stored or logged.",
              "title": "X-Gemini-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            },
            "description": "SSE events: start, node, result, error"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or missing Gemini API key (live mode)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited (this API or the Gemini key's quota)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Generate an assessment, streaming per-step progress (Server-Sent Events)",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{assessment_id}": {
      "delete": {
        "operationId": "delete_assessment_api_v1_assessments__assessment_id__delete",
        "parameters": [
          {
            "description": "Assessment UUID",
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "description": "Assessment UUID",
              "maxLength": 36,
              "minLength": 36,
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or missing Gemini API key (live mode)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited (this API or the Gemini key's quota)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Delete Assessment",
        "tags": [
          "assessments"
        ]
      },
      "get": {
        "operationId": "get_assessment_api_v1_assessments__assessment_id__get",
        "parameters": [
          {
            "description": "Assessment UUID",
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "description": "Assessment UUID",
              "maxLength": 36,
              "minLength": 36,
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assessment"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or missing Gemini API key (live mode)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited (this API or the Gemini key's quota)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Assessment",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{assessment_id}/export.{fmt}": {
      "get": {
        "operationId": "export_api_v1_assessments__assessment_id__export__fmt__get",
        "parameters": [
          {
            "description": "Assessment UUID",
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "description": "Assessment UUID",
              "maxLength": 36,
              "minLength": 36,
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "fmt",
            "required": true,
            "schema": {
              "enum": [
                "pdf",
                "ics",
                "json"
              ],
              "title": "Fmt",
              "type": "string"
            }
          },
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/pdf": {},
              "text/calendar": {}
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or missing Gemini API key (live mode)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited (this API or the Gemini key's quota)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Export as PDF, iCalendar (.ics) or JSON",
        "tags": [
          "assessments"
        ]
      }
    },
...
    "/api/v1/assessments/{assessment_id}/replan": {
      "post": {
        "operationId": "replan_api_v1_assessments__assessment_id__replan_post",
        "parameters": [
          {
            "description": "Assessment UUID",
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "description": "Assessment UUID",
              "maxLength": 36,
              "minLength": 36,
              "title": "Assessment Id",
              "type": "string"
            }
          },
          {
            "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
            "in": "header",
            "name": "x-device-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Anonymous, client-generated identifier that scopes saved plans to one browser.",
              "title": "X-Device-Id"
            }
          },
          {
            "description": "Bring-your-own Gemini API key. Used for this request only; never stored or logged.",
            "in": "header",
            "name": "x-gemini-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bring-your-own Gemini API key. Used for this request only; never stored or logged.",
              "title": "X-Gemini-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assessment"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid or missing Gemini API key (live mode)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Rate limited (this API or the Gemini key's quota)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Regenerate the next training block using logged progress",
        "tags": [
          "assessments"
        ]
      }
    },
...
  }
}